+++ src/main/java/de/tosox/zonerelay/util/ProgressListener.java
@@ -2,5 +2,5 @@ package de.tosox.zonerelay.util;
 
 @FunctionalInterface
 public interface ProgressListener {
-	void onProgressUpdate(int current, int total);
+	void onProgressUpdate(long current, long total);
 }
